refactored input-type folder of user folder
